mito-ai: connect to mito-ai server
